requirements.txt
discord.py
pillow
requests
bs4